Last additions for the menu

+++ include/rpg.h
@@ -65,8 +65,8 @@
         parallax_t *font_cpy;
         npc_t *npc;
         music_t *music;
-        float music_volume;
-        float sound_volume;
+        int music_volume;
+        int sound_volume;
         char *save_path;
     } rpg_t;
